requirements.txt
aiohttp==3.7.3
astroid==2.14.2
async-timeout==3.0.1
attrs==20.3.0
black==23.1.0
chardet==3.0.4
click==8.1.3
colorama==0.4.6
dill==0.3.6
discord.py==1.6.0
idna==3.1
isort==5.12.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
multidict==5.1.0
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.0
platformdirs==3.1.0
pylint==2.16.3
python-dotenv==0.15.0
tomli==2.0.1
tomlkit==0.11.6
typing-extensions==4.5.0
wrapt==1.15.0
yarl==1.6.3
